Use a token-based approach to backend selection.
This change updates Pool to support arbitrary costs of picking an entry, by
providing a reference-counting Token. The connection-counting logic is moved in
a generic fashion into Pool.

The Backend/Balancer types are updated with a new BackendToken. It is now
required to obtain a BackendToken before accepting an incoming TPC connection.

The loop package now operates more generically.

Connection maintenance is now smeared over time.